# api/docs/swagger.yaml
definitions:
...
  models.ChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  models.UserUpdate:
    properties:
      email:
        type: string
      full_name:
        type: string
      phone_number:
        type: string
    type: object
...
  user.Profile:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server for Api-gateway Service
  title: Api Gateway
  version: "1.0"
paths:
...
  /user/password:
    put:
      description: Changes user password
      parameters:
      - description: Passwords
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Changes password
      tags:
      - user
  /user/profile:
    get:
      description: Retrieves user profile
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Profile'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets profile
      tags:
      - user
  /user/profile/delete:
    delete:
      description: Deletes user profile
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Deletes profile
      tags:
      - user
  /user/profile/update:
    put:
      description: Updates user profile
      parameters:
      - description: New user data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdate'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates profile
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"